lock

+++ lock-xss.sh
@@ -6,9 +6,21 @@
 ## CONFIGURATION ##############################################################
 
 
+TMPBG=/tmp/screen.png
+
+B='#00000000'  # blank
+C='#ffffff22'  # clear ish
+D='#689d68cc'  # default
+T='#d79921ff'  # text
+W='#cc241dbb'  # wrong
+V='#b16286bb'  # verifying
+
+i3lock_options="-i $TMPBG --insidevercolor=$C --ringvercolor=$V --insidewrongcolor=$C --ringwrongcolor=$W --insidecolor=$B --ringcolor=$D --linecolor=$B --separatorcolor=$D --verifcolor=$T --wrongcolor=$T --timecolor=$T --datecolor=$T --layoutcolor=$T --keyhlcolor=$W --bshlcolor=$W --screen 1 --clock --indicator --timestr=%H:%M:%S --datestr=%A,%m,%Y --keylayout 2"
+
 # Run before starting the locker
 pre_lock() {
     xinput --disable $(xinput --list | sed -rn 's/.*Mouse.*Mouse.*id=([0-9]+).*/\1/p')
+    scrot $TMPBG && convert $TMPBG -scale 5% -scale 2000% $TMPBG
     #mpc pause
     return
 }
@@ -43,9 +55,12 @@ if [[ -e /dev/fd/${XSS_SLEEP_LOCK_FD:--1} ]]; then
         sleep 0.5
     done
 else
-    trap 'kill %%' TERM INT
-    i3lock -n $i3lock_options &
-    wait
+    eval $(xdotool getmouselocation --shell)
+    if [ "$X" -gt 30 ] || [ "$1" != "" ] ; then
+      trap 'kill %%' TERM INT
+      i3lock -n $i3lock_options &
+      wait
+    fi
 fi
 
 post_lock
